# contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ColumnDto": {
        "description": "Опис колонки для клієнта.",
        "properties": {
          "code": {
            "type": "string"
          },
          "dataType": {
            "type": "string"
          },
          "defaultValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "displayFormat": {
            "type": [
              "null",
              "string"
            ]
          },
          "header": {
            "type": "string"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "isReadOnly": {
            "type": "boolean"
          },
          "isRequired": {
            "type": "boolean"
          },
          "lookupRegistryDefId": {
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "ordinal": {
            "format": "int32",
            "type": "integer"
          },
          "precision": {
            "format": "uint8",
            "type": [
              "null",
              "integer"
            ]
          },
          "scale": {
            "format": "uint8",
            "type": [
              "null",
              "integer"
            ]
          },
          "unitId": {
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "unitSymbol": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "id",
          "code",
          "header",
          "dataType",
          "ordinal",
          "isReadOnly",
          "isRequired",
          "displayFormat",
          "defaultValue",
          "lookupRegistryDefId",
          "unitId",
          "unitSymbol"
        ],
        "type": "object"
      },
...
      "CreateRowRequest": {
        "description": "Запит на створення рядка.",
        "properties": {
          "rowKey": {
            "description": "Бажаний ключ; `null` — згенерувати GUID.",
            "type": [
              "null",
              "string"
            ]
          },
          "tableInstanceId": {
            "description": "Екземпляр таблиці.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "tableInstanceId",
          "rowKey"
        ],
        "type": "object"
      },
...
      "DocumentSummary": {
        "description": "Документ у переліку.",
        "properties": {
          "businessKey": {
            "description": "Бізнес-ключ, унікальний у межах проєкту.",
            "type": "string"
          },
          "createdAt": {
            "description": "Момент створення.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Ідентифікатор.",
            "format": "int64",
            "type": "integer"
          },
          "projectId": {
            "description": "Проєкт.",
            "format": "int32",
            "type": "integer"
          },
          "sheetCount": {
            "description": "Скільки аркушів у складі.",
            "format": "int32",
            "type": "integer"
          },
          "sheetStates": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Стан робочого процесу: аркуш → статус.",
            "type": "object"
          }
        },
        "required": [
          "id",
          "projectId",
          "businessKey",
          "createdAt",
          "sheetCount",
          "sheetStates"
        ],
        "type": "object"
      },
...
      "PatchCell": {
        "description": "Зміна однієї комірки. Три різні операції (R-B4):\r\nзначення — записати; `Value = null` — стерти (рядок видаляється);\r\n`IsEmpty = true` — явна порожнеча; поле відсутнє в запиті — не чіпати.",
        "properties": {
          "columnCode": {
            "description": "Код колонки.",
            "type": "string"
          },
          "isEmpty": {
            "default": false,
            "description": "Явна порожнеча.",
            "type": "boolean"
          },
          "value": {
            "description": "Значення; `null` = стерти."
          }
        },
        "required": [
          "columnCode",
          "value"
        ],
        "type": "object"
      },
      "PatchCellsRequest": {
        "description": "Пакетна зміна комірок. Часткове застосування заборонене: конфлікт у\r\nбудь-якому рядку відхиляє весь батч (B04 §2.3).",
        "properties": {
          "origin": {
            "description": "Джерело зміни: `UserEdit`, `Import`, `Recalculation`.",
            "type": "string"
          },
          "periodKey": {
            "description": "Ключ періоду.",
            "format": "int32",
            "type": "integer"
          },
          "rows": {
            "description": "Рядки зі змінами.",
            "items": {
              "$ref": "#/components/schemas/PatchRow"
            },
            "type": "array"
          },
          "tableInstanceId": {
            "description": "Екземпляр таблиці.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "tableInstanceId",
          "periodKey",
          "origin",
          "rows"
        ],
        "type": "object"
      },
      "PatchCellsResponse": {
        "description": "Результат пакетної зміни.",
        "properties": {
          "appliedCells": {
            "description": "Скільки комірок записано.",
            "format": "int32",
            "type": "integer"
          },
          "rowVersions": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Нові версії зачеплених рядків: `RowKey` → hex.",
            "type": "object"
          },
          "validation": {
            "description": "Результати валідації рівнів, які не блокують запис (R-B3).",
            "items": {
              "$ref": "#/components/schemas/ValidationMessageDto"
            },
            "type": "array"
          }
        },
        "required": [
          "appliedCells",
          "rowVersions",
          "validation"
        ],
        "type": "object"
      },
      "PatchRow": {
        "description": "Рядок у пакетній зміні.",
        "properties": {
          "baseVersion": {
            "description": "Версія рядка, від якої відштовхується клієнт (hex `rowversion`).\r\n`null` означає <b>створення</b> нового рядка (R-B2).",
            "type": [
              "null",
              "string"
            ]
          },
          "cells": {
            "description": "Зміни комірок.",
            "items": {
              "$ref": "#/components/schemas/PatchCell"
            },
            "type": "array"
          },
          "rowKey": {
            "description": "Ідентичність рядка.",
            "type": "string"
          }
        },
        "required": [
          "rowKey",
          "baseVersion",
          "cells"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
      "RowDto": {
        "description": "Рядок зі значеннями. Ключ у Cells — код колонки.",
        "properties": {
          "cells": {
            "additionalProperties": {},
            "description": "Значення; ключ — код колонки. Присутній ключ зі значенням\r\n            `null` означає <b>явну порожнечу</b>, відсутній ключ — «не заповнювали» (R-B4).",
            "type": "object"
          },
          "isOrphaned": {
            "default": false,
            "description": "Рядок посилається на запис реєстру, що втратив чинність (ФВ-8.13).\r\n⚠ Читається зі збереженого поля `doc.TableRow.IsOrphaned`, а не\r\nобчислюється при читанні: перерахунок на кожен зріз не вкладається в\r\nбюджет 400 мс. Читання не блокує, `Submit` блокує.",
            "type": "boolean"
          },
          "label": {
            "description": "Підпис для фіксованих рядків.",
            "type": [
              "null",
              "string"
            ]
          },
          "ordinal": {
            "description": "Позиція.",
            "format": "int32",
            "type": "integer"
          },
          "rowKey": {
            "description": "Ідентичність рядка.",
            "type": "string"
          },
          "rowKind": {
            "description": "Режим рядків таблиці.",
            "type": "string"
          },
          "rowVersion": {
            "description": "Версія для оптимістичного блокування.",
            "type": "string"
          }
        },
        "required": [
          "rowKey",
          "ordinal",
          "rowKind",
          "label",
          "rowVersion",
          "cells"
        ],
        "type": "object"
      },
      "RowKeyResponse": {
        "description": "Створений рядок динамічної таблиці.",
        "properties": {
          "rowKey": {
            "description": "Ключ рядка; на нього посилаються формули й аудит.",
            "type": "string"
          }
        },
        "required": [
          "rowKey"
        ],
        "type": "object"
      },
...
      "TableSliceDto": {
        "description": "Зріз таблиці для grid. Порожні комірки не передаються — клієнт бере\r\n`DefaultValue` з опису колонки (ФВ-3.8).\r\nБюджет усієї операції: p95 1.5 с на 500×60 (tz/08 §8.2).",
        "properties": {
          "cellPermissions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnDto"
            },
            "type": "array"
          },
          "periodKey": {
            "format": "int32",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RowDto"
            },
            "type": "array"
          },
          "tableInstanceId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "tableInstanceId",
          "periodKey",
          "columns",
          "rows",
          "cellPermissions"
        ],
        "type": "object"
      },
...
      "ValidationMessageDto": {
        "description": "Повідомлення валідації.",
        "properties": {
          "columnCode": {
            "description": "Колонка; `null` — рівень рядка.",
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "description": "Локалізований текст.",
            "type": "string"
          },
          "rowKey": {
            "description": "Рядок, якого стосується; `null` — рівень таблиці.",
            "type": [
              "null",
              "string"
            ]
          },
          "ruleCode": {
            "description": "Код правила з `cfg.ValidationRule`.",
            "type": "string"
          },
          "severity": {
            "description": "Рівень: `Info`/`Warning`/`Error`.",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "ruleCode",
          "message",
          "rowKey",
          "columnCode"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Ecr.Api | v1"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/documents/{documentId}/cells": {
      "patch": {
        "description": "Часткове застосування заборонене: конфлікт у будь-якому рядку відхиляє\r\nвесь батч із переліком розбіжностей (409 `ECR-CELL-0409`).",
        "parameters": [
          {
            "in": "path",
            "name": "documentId",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCellsRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCellsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCellsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchCellsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchCellsResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PatchCellsResponse"
                }
              }
            },
            "description": "OK"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Пакетна зміна комірок.",
        "tags": [
          "Cells"
        ]
      }
    },
    "/api/v1/documents/{documentId}/rows": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "documentId",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRowRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRowRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowKeyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowKeyResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RowKeyResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Додає рядок у динамічну таблицю.",
        "tags": [
          "Cells"
        ]
      }
    },
    "/api/v1/documents/{documentId}/tables/{tableInstanceId}": {
      "get": {
        "description": "Бюджет: p95 1.5 с на 500×60 (tz/08 §8.2).",
        "parameters": [
          {
            "in": "path",
            "name": "documentId",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "tableInstanceId",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableSliceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableSliceDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TableSliceDto"
                }
              }
            },
            "description": "OK"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "summary": "Зріз таблиці для grid.",
        "tags": [
          "Cells"
        ]
      }
    },
    "/api/v1/documents/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "periodKey",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummary"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummary"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Документ і його аркуші. Право `Document.View`.",
        "tags": [
          "Documents"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}